requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.4
argparse==1.2.1
google-api-python-client==1.5.0
gunicorn==19.4.5
httplib2==0.9.2
itsdangerous==0.24
oauth2client==2.0.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.0
pytz==2016.1
rsa==3.3
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
wsgiref==0.1.2
